Migrate debug config
